+++ backend/tessend/tessend.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"net/http"
 	"strconv"
@@ -14,6 +15,7 @@ import (
 	"github.com/coreos/etcd/clientv3"
 	"github.com/google/uuid"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
+	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/ringv2"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/backend/store/etcd"
@@ -22,8 +24,16 @@ import (
 )
 
 const (
+	// componentName identifies Tessend as the component/daemon implemented in this
+	// package.
+	componentName = "tessend"
+
 	// tessenURL is the http endpoint for the tessen service.
 	tessenURL = "https://tessen.sensu.io/v2/data"
+
+	// tessenIntervalHeader is the name of the header that the tessen service
+	// will return to update the reporting interval of the tessen daemon.
+	tessenIntervalHeader = "tessen-reporting-interval"
 
 	// ringUpdateInterval is the interval, in seconds, that TessenD will
 	// update the ring with any added/removed cluster members.
@@ -36,16 +46,19 @@ const (
 
 // Tessend is the tessen daemon.
 type Tessend struct {
-	interval  uint32
-	store     store.Store
-	ctx       context.Context
-	cancel    context.CancelFunc
-	errChan   chan error
-	ring      *ringv2.Ring
-	interrupt chan *corev2.TessenConfig
-	client    *clientv3.Client
-	url       string
-	backendID string
+	interval     uint32
+	store        store.Store
+	ctx          context.Context
+	cancel       context.CancelFunc
+	errChan      chan error
+	ring         *ringv2.Ring
+	interrupt    chan *corev2.TessenConfig
+	client       *clientv3.Client
+	url          string
+	backendID    string
+	bus          messaging.MessageBus
+	messageChan  chan interface{}
+	subscription messaging.Subscription
 }
 
 // Option is a functional option.
@@ -56,17 +69,20 @@ type Config struct {
 	Store    store.Store
 	RingPool *ringv2.Pool
 	Client   *clientv3.Client
+	Bus      messaging.MessageBus
 }
 
 // New creates a new TessenD.
 func New(c Config, opts ...Option) (*Tessend, error) {
 	t := &Tessend{
-		interval:  corev2.DefaultTessenInterval,
-		store:     c.Store,
-		client:    c.Client,
-		errChan:   make(chan error, 1),
-		url:       tessenURL,
-		backendID: uuid.New().String(),
+		interval:    corev2.DefaultTessenInterval,
+		store:       c.Store,
+		client:      c.Client,
+		errChan:     make(chan error, 1),
+		url:         tessenURL,
+		backendID:   uuid.New().String(),
+		bus:         c.Bus,
+		messageChan: make(chan interface{}, 1),
 	}
 	t.ctx, t.cancel = context.WithCancel(context.Background())
 	t.interrupt = make(chan *corev2.TessenConfig, 1)
@@ -93,6 +109,13 @@ func (t *Tessend) Start() error {
 		return err
 	}
 
+	sub, err := t.bus.Subscribe(messaging.TopicTessen, componentName, t)
+	t.subscription = sub
+	if err != nil {
+		return err
+	}
+
+	go t.startMessageHandler()
 	go t.startWatcher()
 	go t.startRingUpdates()
 	go t.start(tessen)
@@ -111,8 +134,11 @@ func (t *Tessend) Stop() error {
 	} else {
 		logger.WithField("key", t.backendID).Debug("removed a key from the ring")
 	}
+	if err := t.subscription.Cancel(); err != nil {
+		logger.WithError(err).Error("unable to unsubscribe from message bus")
+	}
 	t.cancel()
-	close(t.errChan)
+	close(t.messageChan)
 	return nil
 }
 
@@ -123,7 +149,38 @@ func (t *Tessend) Err() <-chan error {
 
 // Name returns the daemon name.
 func (t *Tessend) Name() string {
-	return "tessend"
+	return componentName
+}
+
+// Receiver returns the tessen receiver channel.
+func (t *Tessend) Receiver() chan<- interface{} {
+	return t.messageChan
+}
+
+func (t *Tessend) startMessageHandler() {
+	for {
+		msg, ok := <-t.messageChan
+		if !ok {
+			logger.Debug("tessen message channel closed")
+			return
+		}
+
+		tessen, ok := msg.(*corev2.TessenConfig)
+		if !ok {
+			logger.WithField("msg", msg).Errorf("received non-TessenConfig on tessen message channel")
+			return
+		}
+
+		data := t.getDataPayload()
+		t.getTessenConfigMetrics(time.Now().UTC().Unix(), tessen, data)
+		logger.WithFields(logrus.Fields{
+			"url":                       t.url,
+			"id":                        data.Cluster.ID,
+			"opt-out":                   tessen.OptOut,
+			data.Metrics.Points[0].Name: data.Metrics.Points[0].Value,
+		}).Info("sending opt-out status event to tessen")
+		_ = t.send(data)
+	}
 }
 
 // startWatcher watches the TessenConfig store for changes to the opt-out configuration.
@@ -264,7 +321,8 @@ func (t *Tessend) enabled(tessen *corev2.TessenConfig) bool {
 // Errors are logged and tessen continues to the best of its ability.
 func (t *Tessend) collectAndSend(tessen *corev2.TessenConfig) {
 	// collect data
-	data := t.collect(time.Now().UTC().Unix())
+	data := t.getDataPayload()
+	t.getTessenMetrics(time.Now().UTC().Unix(), data)
 
 	logger.WithFields(logrus.Fields{
 		"url":                       t.url,
@@ -274,19 +332,14 @@ func (t *Tessend) collectAndSend(tessen *corev2.TessenConfig) {
 	}).Info("sending data to tessen")
 
 	// send data
-	resp, err := t.send(data)
-	if err != nil {
-		logger.WithError(err).Error("tessen phone-home service failed")
-		return
-	}
-	if resp.StatusCode >= 400 {
-		body, _ := ioutil.ReadAll(resp.Body)
-		logger.Errorf("bad status: %d (%q)", resp.StatusCode, string(body))
+	respHeader := t.send(data)
+	if respHeader == "" {
+		logger.Debug("no tessen response header")
 		return
 	}
 
 	// parse the response header for an integer value
-	interval, err := strconv.ParseUint(resp.Header.Get("tessen-reporting-interval"), 10, 32)
+	interval, err := strconv.ParseUint(respHeader, 10, 32)
 	if err != nil {
 		logger.Debugf("invalid tessen response header: %v", err)
 		return
@@ -307,35 +360,25 @@ func (t *Tessend) collectAndSend(tessen *corev2.TessenConfig) {
 	}
 }
 
-// collect data and populate the data payload
-func (t *Tessend) collect(now int64) *Data {
+// getDataPayload retrieves cluster, version, and license information
+// and returns the populated data payload.
+func (t *Tessend) getDataPayload() *Data {
 	var clusterID string
-	var entityCount, backendCount float64
 
-	// collect client count
-	entities, err := t.store.GetEntities(t.ctx, &store.SelectionPredicate{})
+	// collect cluster id
+	cluster, err := t.client.Cluster.MemberList(t.ctx)
 	if err != nil {
-		logger.WithError(err).Error("unable to retrieve client count")
+		logger.WithError(err).Error("unable to retrieve cluster id")
 	}
-	if entities != nil {
-		entityCount = float64(len(entities))
-	}
-
-	// collect server count and cluster id
-	servers, err := t.client.Cluster.MemberList(t.ctx)
-	if err != nil {
-		logger.WithError(err).Error("unable to retrieve cluster information")
-	}
-	if servers != nil {
-		clusterID = fmt.Sprintf("%x", servers.Header.ClusterId)
-		backendCount = float64(len(servers.Members))
+	if cluster != nil {
+		clusterID = fmt.Sprintf("%x", cluster.Header.ClusterId)
 	}
 
 	// collect license information
 	wrapper := &Wrapper{}
 	err = etcd.Get(t.ctx, t.client, licenseStorePath, wrapper)
 	if err != nil {
-		logger.Debugf("cannot retrieve license: %v", err)
+		logger.WithError(err).Debug("unable to retrieve license")
 	}
 
 	// populate data payload
@@ -345,27 +388,82 @@ func (t *Tessend) collect(now int64) *Data {
 			Version: version.Semver(),
 			License: wrapper.Value.License,
 		},
-		Metrics: corev2.Metrics{
-			Points: []*corev2.MetricPoint{
-				&corev2.MetricPoint{
-					Name:      "entity_count",
-					Value:     entityCount,
-					Timestamp: now,
-				},
-				&corev2.MetricPoint{
-					Name:      "backend_count",
-					Value:     backendCount,
-					Timestamp: now,
-				},
-			},
-		},
 	}
 
 	return data
 }
 
-// send the data payload to the tessen url
-func (t *Tessend) send(data *Data) (*http.Response, error) {
+// getTessenMetrics populates the data payload with # backends/entities.
+func (t *Tessend) getTessenMetrics(now int64, data *Data) {
+	var entityCount, backendCount float64
+
+	// collect entity count
+	entities, err := etcd.Count(t.ctx, t.client, etcd.GetEntitiesPath(t.ctx, ""))
+	if err != nil {
+		logger.WithError(err).Error("unable to retrieve entity count")
+	}
+	entityCount = float64(entities)
+
+	// collect backend count
+	cluster, err := t.client.Cluster.MemberList(t.ctx)
+	if err != nil {
+		logger.WithError(err).Error("unable to retrieve backend count")
+	}
+	if cluster != nil {
+		backendCount = float64(len(cluster.Members))
+	}
+
+	// populate data payload
+	data.Metrics = corev2.Metrics{
+		Points: []*corev2.MetricPoint{
+			&corev2.MetricPoint{
+				Name:      "entity_count",
+				Value:     entityCount,
+				Timestamp: now,
+			},
+			&corev2.MetricPoint{
+				Name:      "backend_count",
+				Value:     backendCount,
+				Timestamp: now,
+			},
+		},
+	}
+}
+
+// getTessenConfigMetrics populates the data payload with an opt-out status event.
+func (t *Tessend) getTessenConfigMetrics(now int64, tessen *corev2.TessenConfig, data *Data) {
+	data.Metrics = corev2.Metrics{
+		Points: []*corev2.MetricPoint{
+			&corev2.MetricPoint{
+				Name:      "tessen_config_update",
+				Value:     1,
+				Timestamp: now,
+				Tags: []*corev2.MetricTag{
+					&corev2.MetricTag{
+						Name:  "opt_out",
+						Value: strconv.FormatBool(tessen.OptOut),
+					},
+				},
+			},
+		},
+	}
+}
+
+// send sends the data payload to the tessen url and retrieves the interval response header.
+func (t *Tessend) send(data *Data) string {
 	b, _ := json.Marshal(data)
-	return http.Post(t.url, "application/json", bytes.NewBuffer(b))
+	resp, err := http.Post(t.url, "application/json", bytes.NewBuffer(b))
+	// TODO(nikki): special case logs on a per error basis
+	if err != nil {
+		logger.WithError(err).Error("tessen phone-home service failed")
+		return ""
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode >= 400 {
+		body, _ := ioutil.ReadAll(io.LimitReader(resp.Body, 4096))
+		logger.Errorf("bad status: %d (%q)", resp.StatusCode, string(body))
+		return ""
+	}
+
+	return resp.Header.Get(tessenIntervalHeader)
 }