+++ dashboard/src/app/component/partial/SilenceEntryForm/index.js
@@ -1,2 +1,0 @@
-export { default as SilenceEntryForm } from "./SilenceEntryForm";
-export { default as SilenceEntryFormFields } from "./SilenceEntryFormFields";