+++ dashboard/src/lib/component/base/QuickNav/index.js
@@ -1,2 +1,0 @@
-export { default } from "./Nav";
-export { default as Button } from "./Button";