+++ dashboard/src/lib/component/util/WithWidth/index.js
@@ -1,6 +1,0 @@
-export { default } from "./WithWidth";
-export {
-  default as withWidth,
-  isWidthUp,
-  isWidthDown,
-} from "@material-ui/core/withWidth";