+++ dashboard/src/lib/constant/polling.js
@@ -1,3 +1,0 @@
-export const pollingDuration = {
-  short: 5000, // 5s
-};