+++ dashboard/src/lib/util/buildInfo.js
@@ -1,1 +1,0 @@
-export { sourceRevision, sourceVersion, sourceURL } from "./buildInfo.macro";