+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Added
 - Added the message bus to Tessend in order to track Tessen configuration changes from the API.
 - Added a performance optimizing `Count()` function to the generic store.
+- Added a hexadecimal Cluster ID title to the `sensuctl cluster health` and
+`sensuctl cluster member-list` commands in tabular format.
+- Added a `Header` field to the `HealthResponse` type returned by `/health`.
 
 ## [5.5.1] - 2019-04-15