+++ CHANGELOG.md
@@ -5,6 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.6.0] - TBD
+
+### Added
+- Added fields getter functions for resources available via the REST API.
+
+### Fixed
+- Fixed the agent `--annotations` and `--labels` flags.
+
+### Added
+- Added the message bus to Tessend in order to track Tessen configuration changes from the API.
+- Added a performance optimizing `Count()` function to the generic store.
+- Added a hexadecimal Cluster ID title to the `sensuctl cluster health` and
+`sensuctl cluster member-list` commands in tabular format.
+- Added a `Header` field to the `HealthResponse` type returned by `/health`.
+
 ## [5.5.1] - 2019-04-15
 
 ### Changed